moved demo data into separate file
